Rename jax_utils into global_setting
