+++ dynamiqs/utils/__init__.py
@@ -1,4 +1,4 @@
-from .jax_utils import *
+from .global_settings import *
 from .operators import *
 from .optimal_control import *
 from .quantum_utils import *
