+++ dynamiqs/utils/global_settings.py
