+++ mkdocs.yml
@@ -230,11 +230,11 @@ nav:
           - QArray utilities:
               - stack: python_api/qarrays/utils/stack.md
               - set_layout: python_api/qarrays/utils/set_layout.md
-          - JAX-related utilities:
-              - to_qutip: python_api/utils/jax_utils/to_qutip.md
-              - set_device: python_api/utils/jax_utils/set_device.md
-              - set_precision: python_api/utils/jax_utils/set_precision.md
-              - set_matmul_precision: python_api/utils/jax_utils/set_matmul_precision.md
+          - Global settings:
+              - to_qutip: python_api/utils/global_settings/to_qutip.md
+              - set_device: python_api/utils/global_settings/set_device.md
+              - set_precision: python_api/utils/global_settings/set_precision.md
+              - set_matmul_precision: python_api/utils/global_settings/set_matmul_precision.md
           - Vectorization:
               - operator_to_vector: python_api/utils/vectorization/operator_to_vector.md
               - vector_to_operator: python_api/utils/vectorization/vector_to_operator.md
